fix: prevent int overflow when calculating range size (#37)
Update the code calculating the size of a range to use a `long` variable
instead of an `int` to avoid overflow errors when selecting extremely large
ranges (e.g. an entire sheet).
To ensure consistency, update all variables dealing with range sizes to use
a `long` type instead of `int`.

BREAKING CHANGE: All references to range size now stored as long variables instead of int

+++ ExcelChangeHandler/Api/Config/IConfiguration.cs
@@ -18,6 +18,6 @@ namespace ExcelChangeHandler.Api.Config
         /// This means that the entire range will be flagged as changed.
         /// Setting this to a high value may cause out-of-memory errors.
         /// </summary>
-        int MaxMemorySize { get; set; }
+        long MaxMemorySize { get; set; }
     }
 }
